Correct type for Channel#perform
With `skipLibCheck` omitted, Channel#perform would fail tsc with:

```
Type 'ActionsType[A]' does not satisfy the constraint '(...args: any) => any'.
  Type 'ActionsType[keyof ActionsType]' is not assignable to type '(...args: any) => any'.
    Type 'ActionsType[string] | ActionsType[number] | ActionsType[symbol]' is not assignable to type '(...args: any) => any'.
```

This is because `Parameters` expects to take in a function, but the
existing ternary doesn't ensure that it's being called on a function.

Parameters: https://www.typescriptlang.org/docs/handbook/utility-types.html#parameterstype

So we should use the `ActionsType[A] extends () => void` to guard
for a function, but that was too strict because it would only match
against functions that took _no_ parameters.

If we expand the type to take any function with any amount of
parameters, then we can safely call `Parameters` on `ActionsType`.
And we can use the `[]` at the very end to accept any other type
`ActionsType` might be.

Co-authored-by: throwandgo <7328768+throwandgo@users.noreply.github.com>