Add type check for declaration files in `yarn test`
It looks like in tsconfig.json, the `skipLibCheck` flag was added
so that check-dts could ignore these external errors:

```
✖ node_modules/@types/node/globals.d.ts:67:13: Type error TS2502
  'AbortController' is referenced directly or indirectly in its own type annotation.

✖ node_modules/@types/node/globals.d.ts:74:13: Type error TS2502
  'AbortSignal' is referenced directly or indirectly in its own type annotation.
```

See the check-dts FAQ for their guidance:
https://github.com/ai/check-dts/blob/7020f61159798fbe19d2a1e38aadd75ab0c7ccc5/README.md#i-am-getting-an-error-from-node-types-how-do-i-skip-node_modules

But adding that flag meant the type declaration files (*.d.ts) in
this anycable-client repo were skipped, so any errors were
overlooked. This has downstream effects for any consumers of the
anycable-core or anycable-web packages. If any declaration files
had type errors (as they did), and the consumers did not set
`skipLibCheck` in their tsconfig, then they'd experience
a failure in their repo as evidenced in issue #39.

So it's prudent for us to _not_ skip checking the type declarations.

`tsc` and `check-dts` both read from tsconfig.json, but we can point
`tsc` to a different configuration file that omits `skipLibCheck`
and ensures this repo's declaration files are checked.

Added the tsc type check at the front because it's the quickest test,
so tests will fail early and fast if there's an error.